Revert "connexion ok / interface a finir"
