+++ src/components/Login.tsx
@@ -1,0 +1,48 @@
+// ./src/components/Login.tsx
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { login } from '../features/auth/authSlice';
+
+const Login: React.FC = () => {
+  const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    dispatch(login({ email, password })).then(() => {
+      navigate('/dashboard');
+    });
+  };
+
+  return (
+    <div className="flex items-center justify-center min-h-screen bg-gray-100">
+      <form className="p-6 bg-white rounded shadow-md" onSubmit={handleSubmit}>
+        <h2 className="text-2xl font-bold mb-4">Connexion</h2>
+        <input
+          type="email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          placeholder="Email"
+          required
+          className="mb-4 p-2 border border-gray-300 rounded w-full"
+        />
+        <input
+          type="password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          placeholder="Mot de passe"
+          required
+          className="mb-4 p-2 border border-gray-300 rounded w-full"
+        />
+        <button type="submit" className="w-full bg-blue-500 text-white p-2 rounded">
+          Connexion
+        </button>
+      </form>
+    </div>
+  );
+};
+
+export default Login;
