--- src/components/pages/Dashboard.tsx
@@ -1,14 +1,0 @@
-// src/components/pages/Dashboard.tsx
-
-import React from 'react';
-
-const Dashboard: React.FC = () => {
-    return (
-        <div className="dashboard">
-            <h2 className="dashboard-title">Tableau de bord</h2>
-            <p className="dashboard-content">Bienvenue sur votre tableau de bord.</p>
-        </div>
-    );
-};
-
-export default Dashboard;
