--- src/styles/auth/AuthForm.css
@@ -1,43 +1,0 @@
-/* src/styles/AuthForm.css */
-.auth-form {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.auth-form-title {
-    text-align: center;
-    font-size: 1.5rem;
-    color: #333;
-}
-
-.auth-form-group {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.auth-form-label {
-    font-size: 1rem;
-    color: #333;
-}
-
-.auth-form-input {
-    padding: 0.5rem;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-}
-
-.auth-form-button {
-    padding: 0.75rem;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.auth-form-button:hover {
-    background-color: #0056b3;
-}
