--- src/styles/pages/AuthPage.css
@@ -1,37 +1,0 @@
-/* src/styles/AuthPage.css */
-.auth-page {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    min-height: 100vh;
-    background-color: #f0f4f8;
-}
-
-.auth-nav {
-    margin-bottom: 2rem;
-    display: flex;
-    gap: 1rem;
-}
-
-.auth-link {
-    padding: 0.5rem 1rem;
-    background-color: #007bff;
-    color: white;
-    border-radius: 4px;
-    text-decoration: none;
-    transition: background-color 0.3s ease;
-}
-
-.auth-link:hover {
-    background-color: #0056b3;
-}
-
-.auth-form-container {
-    width: 100%;
-    max-width: 400px;
-    background-color: white;
-    padding: 2rem;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
