--- src/styles/pages/Dashboard.css
@@ -1,21 +1,0 @@
-/* src/styles/pages/Dashboard.css */
-.dashboard {
-    max-width: 800px;
-    margin: 2rem auto;
-    padding: 2rem;
-    background-color: white;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.dashboard-title {
-    text-align: center;
-    font-size: 2rem;
-    color: #333;
-    margin-bottom: 1rem;
-}
-
-.dashboard-content {
-    font-size: 1.125rem;
-    color: #666;
-}
